requirements.txt
aiogram==3.18.0
beautifulsoup4==4.13.4
crewai==0.86.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
tenacity==9.0.0
requests==2.32.3
PyYAML==6.0.2
setuptools==75.8.0